Simplify ImplicitModelFollower class
The QR-weighted solution is really close to the exact solution, so we
can use the latter and save a lot of compilation time by not using the
DARE solver.
